User registration function implemented once. impot warning fix
